+++ src/contexts/index.js
@@ -1,0 +1,6 @@
+import { useAuth } from "./auth-context";
+import { useCart } from "./cart-context";
+import { useProduct } from "./product-context";
+import { useWishlist } from "./wishlist-context";
+
+export {useAuth, useCart, useProduct, useWishlist}
