+++ src/contexts/wishlist-context.js
@@ -2,6 +2,7 @@ import axios from "axios";
 import { createContext, useContext, useReducer } from "react";
 import { useNavigate } from "react-router-dom";
 import {useAuth} from "./auth-context"
+
 import {wishlistReducer} from "../reducer/wishlistReducer"
 const wishlistContext = createContext()
